X and O position calculation
 0 | 1 | 2
 ---------
  3 | 4 | 5
 ---------
  6 | 7 | 8

For x cordinate transform
(selectedValue % 3) * tileSpace + 10 * (selectedValue  % 3)

For y cordinate transform
(selectedValue  / 3) * tileSpace + 10 * (selectedValue  / 3);

selectedValue refers to the position of the tile on the board ranging from 0 to 8.
tileSpace is the space between each tile on the board.
The addition of 10... is to account for the thickness of the board lines